+++ src/girocheckout-sdk/api/creditcard/GiroCheckout_SDK_CreditCardCapture.php
@@ -37,13 +37,13 @@ class GiroCheckout_SDK_CreditCardCapture extends GiroCheckout_SDK_AbstractApi im
     protected $responseFields = array(
         'rc'=> TRUE,
         'msg' => TRUE,
-        'reference' => TRUE,
-        'referenceParent' => TRUE,
-        'merchantTxId' => TRUE,
-        'backendTxId' => TRUE,
-        'amount' => TRUE,
-        'currency' => TRUE,
-        'resultPayment' => TRUE,
+        'reference' => FALSE,
+        'referenceParent' => FALSE,
+        'merchantTxId' => FALSE,
+        'backendTxId' => FALSE,
+        'amount' => FALSE,
+        'currency' => FALSE,
+        'resultPayment' => FALSE,
     );
 
     /*