+++ src/girocheckout-sdk/api/maestro/GiroCheckout_SDK_MaestroRefund.php
@@ -37,13 +37,13 @@ class GiroCheckout_SDK_MaestroRefund extends GiroCheckout_SDK_AbstractApi implem
     protected $responseFields = array(
         'rc'=> TRUE,
         'msg' => TRUE,
-        'reference' => TRUE,
+        'reference' => FALSE,
         'referenceParent' => FALSE,
-        'merchantTxId' => TRUE,
-        'backendTxId' => TRUE,
-        'amount' => TRUE,
-        'currency' => TRUE,
-        'resultPayment' => TRUE,
+        'merchantTxId' => FALSE,
+        'backendTxId' => FALSE,
+        'amount' => FALSE,
+        'currency' => FALSE,
+        'resultPayment' => FALSE,
     );
 
     /*